Big bug fixes for the 'MMQ 2-site' dispersion model.
The equations in Korzhnev et al., 2004 are incorrect!  This is clear when comparing the equations in
the paper to those implemented in the cpmg_fit source code.  The equations in relax have been
changed to match the correct one.

In addition, the matrix power factor must be found with the python math.floor() function and not
int() as the later will sometimes round up.