Interpolated curves are now produced for the numeric CPMG-type models.
This if for the relax_disp.plot_disp_curves user function.  The resolution of these are limited to
the frequency of a single CPMG block in the relaxation time period.  Therefore the plots are
produced slightly differently.

To enable this functionality, the new count_exp() and return_relax_times() functions have been added
to the specific_analyses.relax_disp.disp_data module.