Bug fixes for the molecule.delete, residue.delete and spin.delete user functions.
The molecule, residue, and spin metadata in the relax data store was not being updated correctly
after these user function calls so that any subsequent operations on this data was failing.  This
metadata problem was not noticed before as it disappears if the state is saved and reloaded into
relax after a restart.