A big change to the usage of the relax_disp.r2eff_read_spin user function.
Now the nu_CPMG frequency or the spin-lock field strength must be set prior to calling this user
function.  This allows for more flexibility as often the experiment IDs and frequency values in the
files do not match to the same number of decimal places.  The frequency is no longer read from the
file but must be preset.