Redesign of the CPMG frequency and spin-lock field strength data structures.
These now have an extra dimension for the offset so that the values are now experiment, magnetic
field strength and offset dependent.  If many offsets are present but are variable for each
dispersion point, then this saves a lot of calculation time.  This mainly affects R1rho-type data.
To better handle this, all of the specific_analyses.relax_disp.disp_data.loop_*() functions have
been modified to accept data values rather than indices.