Large speed up for the internal structural object for when many models are present.
The new ModelList.current_models object keeps track of all the models already present in the
structural object.  This simplifies the checks of the pack_structs() internal structural object
method by removing expensive looping.  This allows the loading of PDB files to continue to be fast
even with many tens or hundreds of thousands of models already loaded.