Created the RDC and PCS back calculated test data for the CaM double rotor frame order model.
This consists of a uniform distribution over both rotors, the first centred in the C-domain and
the second in the N-domain, and the two axes being perpendicular to each other along the CoM-CoM
axis.  The distribution consists of 250000 rotated structures.  The frame order matrix for this
model is also given.