Added the truncated data for creating a system test to catch bug #21562.
This bug (https://gna.org/bugs/?21562) was reported by Dhanas Muthu (https://gna.org/users/dhanas)
and is the failure of the NOE analysis when spectra are replicated.

This consists of the Sparky peak lists attached to the bug report and the modified 2AT7 PDB file.
The data has been truncated to only include residues :12, :13, and :14.