Created the Noe.test_bug_21562_noe_replicate_fail system test.
This is to catch bug #21562 (https://gna.org/bugs/?21562), reported by Dhanas Muthu
(https://gna.org/users/dhanas).  This is the failure of the NOE analysis when spectra are
replicated.

This uses the truncated data taken from the files attached to the bug report.  The NOE output file
is checked to see if the contents are correct.