Improvements for the steady-state NOE analysis overfit_deselect() method.
The spin deselection which occurs at the start of the calc user function call, used to calculate the
NOE, is now clearer.  Each deselection condition is now explained in detail and the text is now far
more informative.

In addition, the special condition of all spins being deselected is now caught.  If this happens, a
RelaxError is raised to prevent the user from going forwards.  This should remove confusion as to
why the output file is empty.