Updated the N_state_model.5_state_xz system test to allow it to complete on i586 Linux systems.
The optimisation would continue for a huge amount of time on a test system (Mageia 4 i586 VM) and
would make it appear as though the test suite has hung.  By limiting the maximum number of
iterations in the optimisation to 1000, the test will complete successfully and the parameters
optimised to the same precision.