Fix for bug #21615, the missing data dialog failure when executing the GUI model-free analysis.
This is the bug at https://gna.org/bugs/?21615 as reported by Ivan Leung (ivanhoe dott leung att
chem dott ox dot ac dot uk).

The problem is that the spin container's "isotope" variable is being accessed directly after a test
showing that this variable does not exist.  This is now fixed so that the missing data dialog is now
presented explaining that the spin isotope information is not set.