Implemented the GUI element for loading multiple files.
This follows from the thread http://thread.gmane.org/gmane.science.nmr.relax.devel/5057/focus=5062.

This is via the new user function argument type "file sel multi", now used by the
spectrum.read_intensities user function.  The file selection element consists of two parts.  The GUI
element embedded in the user function wizard page is similar to the "file sel" element, except that
the preview button is not present.  The file selection button behaviour is also different in that it
launches the new multiple file selection window.

The multiple file selection window is based on the 'sequence' data window, as used in the spectrum
ID argument for the spectrum.read_intensities user function.  However the ListCtrl element has been
replaced by a custom scrolled panel.  The 'Add' button adds a new file selection GUI element
consisting of a TextCtrl for displaying and manual editing of the file name, the file selection
button for launching the relax file selection dialog, and the preview button lost in the parent GUI
element.   The scrolled panel allows more elements in the panel than can fit in the window.  The
'Delete all' and 'OK' buttons from the 'sequence' data window are also present and function as
expected.