Improved the behaviour of the multiple file selection GUI window.
The RelaxFileDialog GUI element is now initialised when the file selection button is clicked rather
than in the __init__() method.  The result of this change is that the current working directory is
dynamically changed in the RelaxFileDialog, hence if the directory is changed in one file selection
element, it will look like it is changed in all.