Fix for bug #21716 (https://gna.org/bugs/?21716).
This is the failure to save the relax state just after deleting the current data pipe, even if other
data pipes exist.

The problem was that the specific analysis functions data_names() and return_data_desc() were being
retrieved using the current data pipe rather than the actual data pipe that the data structures
belong to.  So if the current data pipe is None, then these fail.  Now the data pipe type is being
passed through all of the to_xml() methods so that the correct data_names() and return_data_desc()
methods are retrieved.