Fix for bug #21754 (https://gna.org/bugs/?21754).
This is the failure of the grace.view user function on MS Windows in opening a file in qtgrace when
the path includes spaces, as reported by Mengjun Xue (mengjun dott xue att mailbox dott tu-berlin
dott de).

The fix was to run the program with the file name in double quotes.