Fixes for the rotor axis alpha angle conversion in the CaM frame order system test base script.
The pivot point is not the point on the rotor axis closest to the reference point (the CoM).
Therefore the closest point is now calculated from the pivot point on the axis and the axis vector.
This closest point is needed for defining the new minimal parameter set for the rotor models.