Converted the N_state_model.test_5_state_xz system test to use a new way to set parameters.
Instead of using pseudo-parameters for the value.set user function such as 'p0', 'p1', etc. for the
probabilities, which are then converted into the 'probs' parameter with the index taken from the
name, instead the index is now directly given.  The value.set user function will need to be modified
to handle this.  The aim is to standardise the parameter list object for the N-state model analysis.