Changed the values of the set argument for the parameter list object.
The 'generic' value has been renamed to 'fixed' and is now for all permanently fixed parameters of
the model - for example the CSA value in a number of analyses.  The default set argument value of
the _add() method has been changed to 'all' so that any parameters registered via that method are
not placed in a special set (unless specified otherwise).