Shifted the data_store.relax_xml module into the relax library as lib.xml.
This module contains a couple of functions which are used for converting Python objects into an XML
representation and back again.  These are used not only by the relax data store, but also a number
of the structural objects in the relax library (which are themselves placed in the relax data
store).  This makes the relax library more independent from the rest of relax.