Removed the automatic axis labelling in the lib.software.grace.write_xy_header() function.
This is the automatic labelling based on the parameter Grace strings and units of the specific
analysis type.  This is now preformed by the specific analyses themselves so the automatic code is
not needed or used.

This allows the dependence on the pipe_control.pipes module and the specific analysis API to be
removed making the relax library now 100% independent from the rest of relax.