Renamed the relax_disp.cpmg_frq user function to relax_disp.cpmg_setup and added some new options.
This follows from the thread http://thread.gmane.org/gmane.science.nmr.relax.devel/5511/focus=5520.

The ncyc_even option has been added so the user can specify if the pulse sequence requires an even
number of CPMG blocks.  This is for use in the interpolated dispersion curves, but could have other
uses in the future.