Advances to the Grace 2D plotting abilities in the lib.software.grace relax library module.
The write_xy_header() function now accepts the new 'world', 'tick_major_spacing', and
'tick_minor_count' arguments.  These allow the world view to be preset, and allow the ticks on the
X and Y-axes to be programatically changed.  The write_xy_data() has also been modified so that the
autoscaling can be turned off, as this Grace command will overwrite the world view and tick setup.