Modified the 2nd rotor model of CaM frame order optimisation script.
The frame_order.pdb_model user function is now used to create a PDB representation of the rotor
motions for the real, expected parameters and for the optimisation results when the pivot point is
fixed.

In addition, the pymol.cone_pdb user function has been replaced by the pymol.frame_order user
function.  All new files have been added to the repository.