Deletion of the frame_order.average_position user function and all of the associated backend code.
This user function allowed the user to specify five different types of displacement to the average
moving domain position:

    - A pure rotation, with no translation, about the pivot of the motion in the system.
    - A rotation about the pivot of the motion of the system together with a translation.
    - A pure translation with no rotation.
    - A rotation about the centre of mass of the moving domain with no translation.
    - A rotation about the centre of mass of the moving domain together with a translation.

Now the last option will be the default and only option.  This option is equivalent to the standard
superimposition algorithm (the Kabsch algorithm) to a hypothetical structure at the real average
position.

The other four are due to the history of the development of the theory.  These limit the usefulness
of the theory and will only cause confusion.