Clean up of the frame order target function code.
This matches the previous changes (r23197), the deletion of the frame_order.average_position user
function.  The changes include the removal of the translation optimisation flag as this is now
always performed, and the removal of the flag which causes the average domain rotation pivot point
to match the motional pivot point as these are now permanently decoupled.