Created a script for the CaM frame order test models for finding the average domain position.
As the rotation about a fixed pivot has been eliminated, the shift from 1J7P_1st_NH_rot.pdb to
1J7P_1st_NH.pdb has to be converted into a translation and rotation about the CoM.  This script will
be used to replace the pivot rotation Euler angles with the translation vector and CoM rotation
Euler angles.  However the structure.superimpose user function will need to be modified to handle
both the standard centroid superimposition as well as a CoM superimposition.