Speed-up of model TSMFK01.
task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

This is won by not checking single values in the R2eff array for math domain
errors, but calculating all steps, and in one single round check for finite values.
If just one non-finite value is found, the whole array is returned with a large
penalty of 1e100.

This makes all calculations be the fastest numpy array way.