Math-domain catching for model CR72.
task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

This is to implement catching of math domain errors, before they occur.
These can be found via the --numpy-raise function to the systemtests.

The skipping of test when num_points > 0, is a bad implementation.
If such a case should show, it is best to catch the wrong input for the
calculations. This is best done with a check before running the calculations.