Turned of processor.run_queue() when using calc().
task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

A weird behaviour was found when running the test-suite.

Somehow, queued jobs from the last test suite was not flushed and done.
Running the test-suite gave errors on almost all tests.

Inspecting systemtest which only used calc() for the dispersion API, showed that no jobs was queued
in the calculate() method, and the call to processor.run_queue(), was turned off, solving the problem.