Large increase in accuracy of the RDC and PCS averaging.
This is for the CaM frame order test data generation base script.  By storing the RDCs and PCSs in
numpy.float128 arrays, a 64-bit system is required, and then using numpy.average(), the average
value can be calculated with a much higher accuracy.  As N becomes larger, the averaging introduces
greater and greater amounts of truncation artifacts.  So this change alleviates this.  However a lot
more RAM is now required.