Fix for the chain-reaction failures in the test suite.
This fixes bug #22055 (https://gna.org/bugs/?22055), the processor.run_queue() not cleaning up in
uni_processor - chain-reaction failures in the test suite.

The fix was insanely simple, just implementing what was mentioned Gary's FIXME comment in the
run_queue() method of the uni-processor object.  The queue execution code has been placed inside a
'try' statement and the queue cleaning up code in a 'finally' statement.

This closes a painfully difficult to find bug that has been in relax since 2006.