Large increase in accuracy of the RDC and PCS averaging.
This is for the CaM frame order test data generation base script.  By summing the RDCs and PCSs into
1D numpy.float128 arrays (for this, a 64-bit system is required), and then dividing by N at the end,
the average  value can be calculated with a much higher accuracy.  As N becomes larger, the
numerical averaging introduces greater and greater amounts of truncation artifacts.  So this change
alleviates this.

This change is what r23252 should have been.