Huge increase in precision for the CaM frame order free rotor model test data.
The higher precision is because the number structures in the distribution is now twenty million
rather than one million, and the much higher precision numpy.float128 averaging of the updated data
generation base script has been used.

This data should allow for a much better estimate of the beta and gamma average domain position
parameter values for the free rotor models which are affected by the collapse of the alpha parameter
to zero.