Branch created from trunk to compare to the disp_speed branch.
The branch will be used to write a range of system tests that should
fail in the trunk, and now pass in the disp_speed branch.

This is to investigate:
Bug #22024 (https://gna.org/bugs/index.php?22024) Minimisation space for CR72 is catastrophic. The chi2 surface over dw and pA is bounded.