Modified unit tests demonstrating edge case 'no Rex' failures of the model 'NS CPMG 2-site expanded'.
This is to align with the current return of data in the disp_speed branch.

This follows from the ideas in the post http://article.gmane.org/gmane.science.nmr.relax.devel/5858.
This is related to: task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

This is to implement catching of math domain errors, before they occur.

These tests cover all parameter value combinations which result in no exchange:

    - dw = 0.0,
    - pA = 1.0,
    - kex = 0.0,
    - dw = 0.0 and pA = 1.0,
    - dw = 0.0 and kex = 0.0,
    - pA = 1.0 and kex = 0.0,
    - dw = 0.0, pA = 1.0, and kex = 0.0.