Important bug fix for the structure.delete user function when multiple models are present.
This is to fix bug #22070 (https://gna.org/bugs/index.php?22070), the failure of the
structure.superimpose user function after deleting atoms with structure.delete.

The problem is that structure.delete was removing the atoms from the first model but none of the
others.  This is because it was using the structural object atom_loop() method to find the atoms to
be deleted, but this method operates on the first model.  So when the second model is reached, the
atoms are already gone.