Created the test_ns_cpmg_2site_expanded_no_rex8() relaxation dispersion unit test.
This is a demonstration commit, showing the 'NS CPMG 2-site expanded' model with no exchange when
kex = 1e5.  I.e. when the motion is too fast for exchange to be observed.  This test should be used
for all dispersion models to make sure that they model this edge case correctly as well.

This follows from http://article.gmane.org/gmane.science.nmr.relax.devel/5906.