Attempt at fixing bug #22071, the relax unit test and system test not functioning.
This is reported at https://gna.org/bugs/?22071.

The fix here is that the git commands to show the current subversion revision number only works when
run from the relax base directory, or one of the subdirectories.  This should now be fixed, as the
pipe running the command will first 'cd' to the relax base directory.