Large speed up for the relax start up times for svn and git-svn copies of the relax repository.
The 'svn info' and 'git svn info' commands are now only executed once when the version module is
first imported.  The revision() and url() functions have been merged into the repo_info() function
and this is called when the module is imported.  This repo_info() function stores the repository
revision and URL as the version.repo_revision and version.repo_url module variables.  It also
catches if these variables are already set, so that multiple imports of the module do not cause the
repository information to be looked up each time.

Previously the revision() and url() functions where called every time a relax state or result file
was created, hence for repository copies the 'svn info' or 'git svn info' commands were being called
each time.  The functions were also called for each interpreter object instantiated, and for each
import of the version module.