Updated the CaM frame order free-rotor model test data set for testing for missing data.
This is the data in test_suite/shared_data/frame_order/cam/free_rotor_missing_data.  To simplify the
copying of data from test_suite/shared_data/frame_order/cam/free_rotor and then the deletion of
data, the missing.py script was created to automate the process.  The generate_distribution.py
script and some of the files it creates were removed from the repository so it is clearer how the
data has been created.