Optimisation of the average domain position for the CaM frame order free-rotor models.
The log file that shows the optimisation of the average domain position for the free-rotor models
has been added to the repository for reference.  This is for the simple free-rotor model, but the
optimised position holds for the isotropic cone and pseudo-ellipse model data too.

To perform the optimisation, the axis_theta and axis_phi parameters were removed from the model and
hardcoded into the target function.  As the rotor axis is know, this allows the average domain
position to be optimised in isolation.  Visual inspection of the results confirmed the position to
be correct.