Critical fix for the math domain catching of model TSMFK01.
This was discovered with the added 8 unit tests demonstrating edge case 'no Rex' failures.

This follows from the ideas in the post http://article.gmane.org/gmane.science.nmr.relax.devel/5858.
This is related to: task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

This is to implement catching of math domain errors, before they occur.

The cathing of these occurences needed to be re-written.