Fix for the lib.structure.geometric.generate_vector_residues() function.
The atom numbers are no longer read from the internal structural object, as these are not reliable.
If another geometric representation exists in the object, then the atom numbers could be None.  Or
loading structures from multiple PDB files can cause the numbering to be repeated or out of order.