Fixes for the Sequence and Sequence_2D GUI elements for handling invalid input data.
These elements used by the user function windows previously raised all sorts of errors if the data
was not what they expected (lists or lists of lists respectively).  These situations are now caught
and the input data is ignored, so blank Sequence and Sequence_2D elements are presented to the user.