Added 9th and 10th unit test case for model B14.
These tests are setup, to check how the function behaves under conditions of the grid-seach.

Two tests are setup, which either use the default lower or upper bound of the parameters for the grid search.

The return of the values are not checked, but the behaving of the function can be found when adding the --numpy-raise to run relax.

This is related to: task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.

These tests are implemented to show bug cases related to:
Bug #22032: (bug #22032: Minimisation explodes when using Grid_INC=None) Minimisation explodes when using Grid_INC=None

The function can return nan or inf values, which are not handled by the minimisation algorithm, causing it to stop.
There needs to be mechanism to catch these cases.