Attempt to speed up the pseudo-elliptic frame order models.
The quasi-random numerical integration of the PCS for the pseudo-ellipse has been modified so that
the torsion angle check for each Sobol' point is preformed before the tmax_pseudo_ellipse() function
call.  A new check that the tilt angle is less than cone_theta_y, the larger of the two cone angles,
has also been added to avoid tmax_pseudo_ellipse() when the theta tilt angle is outside of an
isotropic cone defined by cone_theta_y.