Converting back to having back_calc as a function argument to model B14.
This is to clean up the API.
There can be bo no partial measures/implementations in the relax trunk.

The problem is, that many numerical models can't be optimised further, since they
evolve the spin-magnetisation in a matrix.  That spin evolvement can't be put into
a larger numpy array.

This is related to: task #7793: (https://gna.org/task/?7793) Speed-up of dispersion models.