Fix for bug #22112 (http://gna.org/bugs/?22112).
This is the GUI failure when setting list values via the sequence windows, launched from user
function windows fails on Mac OS X.  The problem was two fold.  First the Sequence and Sequence_2D
windows from wx.Dialog should not be terminated via the Destroy() method, as wx.Dialog.Destroy()
appears to be horribly broken on Macs.  Additionally, the setting of data in the Sequence and
Sequence_2D windows is now performed using the thread-safe wx.CallAfter() method.