Proper edge case handling and slight speed up of the frame order PCS integration functions.
The case whereby no Sobol' points in the numerical integration lie within the motional distribution
is now caught and the rotation matrix set to the motional eigenframe to simulate the rigid state.
As the code for averaging the PCS was changed, it was also simplified by removing an unnecessary
loop over all spins.  This should speed up the PCS integration by a tiny amount.