Added one more iteration for the zooming optimisation of the frame order auto-analysis.
This is to improve the speed of optimisation when all RDC and PCS data is being used.  The previous
iterations where with [100, 1000, 200000] Sobol' integration points and [1e-2, 1e-3, 1e-4] function
tolerances.  This has been increased to [100, 1000, 10000, 100000] and [1e-2, 1e-3, 5e-3, 1e-4].
The final number of points has been decreased as that level of accuracy does not appear to be
necessary.  These are also only default values that the user can change for themselves.