relax version 3.2.2.
This is a major feature and bugfix release.  This includes a large speed up of all analytic
relaxation dispersion models, the correct handling of edge case failures in all models of the
dispersion analysis, and a number of fixes for the handling of list-type data in the GUI user
function windows including the fatal GUI crashes on Mac OS X systems.