Expanded the CaM free-rotor isotropic cone test data frame_order.py optimisation script.
The optimisation is now implemented as in the auto-analysis, with an iterative increase in accuracy
of the quasi-random numerical integration together with an decrease of the function tolerance cutoff
for optimisation.  The accuracy of the initial chi-squared calculation is now much higher.  And the
accuracy of the initial grid search and the Monte Carlo simulations is now much lower.

The results of the new optimisation have been added to the repository.