Expanded all remaining CaM test data frame_order.py optimisation scripts.
The optimisation is now implemented as in the auto-analysis, with an iterative increase in accuracy
of the quasi-random numerical integration together with an decrease of the function tolerance cutoff
for optimisation.  The accuracy of the initial chi-squared calculation is now much higher.  And the
accuracy of the initial grid search and the Monte Carlo simulations is now much lower.