Bug fix for r23674, the compilation of the C modules now respects the user defined environment.
This is the patch from Justin (https://gna.org/users/jlec) attached to bug #22145
(https://gna.org/bugs/?22145).

The problem was that on Mac OS X (as well as other systems), that these environmental variables were
not defined and hence the scons commands would all fail with a KeyError and traceback.  Now the keys
in the os.environ dictionary are being searched for before they are set.