Fix for a number of the frame order models which do not have parameter constraints.
This change to the grid_search() API method is similar to the change for the minimise() method
(r23211).  The linear_constraint() function was returning A, b = [], [] for these models, but these
empty numpy arrays were causing the dot product with A to fail in the grid_search() API method.
These values are now caught and the constraint algorithm turned off.