Modified the CaM frame order base system test script to catch a bug in the free rotor model.
The axis spherical angles are no longer set for the rotor or free rotor models, as they use the
alpha angle instead and the lack of the theta and phi parameters triggers the bug.  The PDB
representation of the frame order motions is also now tested for all frame order models, as it was
turned off for the rigid, rotor and free rotor models and this is where the bug lies.