Eliminated the average position alpha Euler angle parameter from the free-rotor pseudo-ellipse model.
As this frame order model is a free-rotor, the average domain position is therefore undefined and it
can freely rotate about the rotor axis.  One of the Euler angles for rotating to the average
position can therefore be removed, just as in the free rotor and free rotor isotropic cone models.