Implemented a masked array search for where "missing" array is equal 1.
This makes it possible to replace all values with this mask, from the value array.

This eliminates the last loops over the missing values.

It took over 4 hours to figure out, that the mask should be called with mask.mask,
to return the same fulls structure,

Task #7807 (https://gna.org/task/index.php?7807): Speed-up of dispersion models for Clustered analysis.