Changed target function for model CR72.
To CR72 is now also the input of the parameters of R20A, R20B and dw.
dw is tested for zero, to return flat lines.
It is faster to search in the smaller numpy array, than the 5 dimensional dw array.

This is for speed-up.

R20A and R20B is also subtracted, to see if the full model should be used.
In the same way, it is faster to subtract the smaller array.

These small tricks are expected to give 5-10 pct. speeed-up.

Task #7807 (https://gna.org/task/index.php?7807): Speed-up of dispersion models for Clustered analysis.