Changed target function for model B14.
To B14 now also send the input of the original parameters dw.
dw is tested for zero, to return flat lines.
It is faster to search in the smaller numpy array, than the 5 dimensional dw array.

This is for speed-up.

These small tricks are expected to give 5-10 pct. speeed-up.

Task #7807 (https://gna.org/task/index.php?7807): Speed-up of dispersion models for Clustered analysis.