Better handling of the control-C keyboard interrupt signal in the relax test suite.
This includes two changes.  The Python 2.7 and higher unittest.installHandler() function is now
called, when present, to terminate all tests using the unittest module control-C handler.  The
second change is that the keyboard interrupt signal is caught in a try-except statement, a message
printed out, and the tests terminated.  This should be an improvement for all systems.