Added a script for timing different ways to calculate PCSs and RDCs for multiple vectors.
This uses the timeit module rather than profile to demonstrate the speed of 7 different ways to
calculate the RDCs or PCSs for an array of vectors using numpy.  In the frame order analysis, this
is the bottleneck for the quasi-random numerical integration of the PCS.

The log file shows a potential 1 order of magnitude speed up between the 1st technique, which is
currently used in the frame order analysis, and the 7th and last technique.  The first technique
loops over each vector, calculating the PCS.  The last expands the PCS/RDC equation of the
projection of the vector into the alignment tensor, and calculates all PCSs simultaneously.