Small speed up for all of the frame order models.
The PCS averaging in the quasi-random numerical integration functions now uses the multiply() and
divide() numpy methods to eliminate a loop over the alignments.  For this, a new dimension over the
spins was added to the PCS constant calculated in the target function __init__() method.  In one
test of the pseudo-ellipse, the time dropped from 191 seconds to 172.