Modified the rigid_test.py system test script to really be the rigid case.
This is used in all of the Frame_order.test_rigid_data_to_*_model system tests.  Previously the
parameters of the dynamics were set to being close to zero, to catch the cases were a few Sobol'
PCS integration points were accepted.  But now the case were no Sobol' points can be used is being
tested.  This checks a code path currently untested in the test suite, demonstrating many failures.