The rigid frame order matrix for the pseudo-ellipse models is now correctly handled.
This allows the rigid case RDCs to be correctly calculated for both the pseudo-ellipse and
torsionless pseudo-ellipse models.  The previous catch of the theta_x cone angle of zero was
incorrectly recreating the frame order matrix, which really should be the identity matrix.  However
truncation artifacts due to the quadratic SciPy integration still cause the model to be
ill-conditioned near the rigid case.  The rigid case is correctly handled, but a tiny shift of the
parameters off zero cause a discontinuity.