Fixes for all of the relaxation dispersion system tests which were failing with the new minfx code.
Due to the tuning of the log barrier constraint algorithm in minfx in the commit at
http://article.gmane.org/gmane.science.mathematics.minfx.scm/25, many system tests needed to be
slightly adjusted.  Two of the Relax_disp.test_tp02_data_to_* system tests were also failing as the
optimisation can no longer move out of the minimum at pA = 0.5 for one spin (due to the low quality
grid search in the auto-analysis).