Modified all of Troels' dispersion profiling scripts to work with older relax versions.
This is in preparation for obtaining some powerful timing statistics.  The calls to the r2eff_*()
functions are unnecessary and are the only failure point in the scripts between the current code in
the disp_spin_speed branch and trunk or older versions of relax.  So these function calls have been
eliminated.