Created a super script for profiling the relaxation dispersion models.
This script will execute all of the current profiling scripts in the directory
test_suite/shared_data/dispersion/profiling for both the current version of relax and any other
specified version (current set to the 3.2.2 relax tag).  It will run the scripts and relax versions
interleaved N=10 times and extract the func_*() target function call profile timings.  This
interleaving makes the numbers much more consistent.  Averages and standard deviations are then
calculated, as well as the speed up between the two relax versions.  The results are printed out in
a format suitable for the relax release messages.