Modified all of the dispersion model profiling scripts.
The single() function for timing the single spin target function speed has been modified to include
a second outer loop over 100 'spins'.  This means that the timing numbers are equivalent to the
cluster timings, as both are then over 100 spins.  This now allows not only relax version
differences and model differences to be compared, but also the non-clustered and clustered analysis
speeds.