Support for the 3 vector system for double motions has been added to the frame order analysis.
This is used for the quasi-random Sobol' numeric integration of the PCS.  The lanthanide to atom
vector is the sum of three parts:  the 1st pivot to atom vector rotated by the 1st mode of motion;
the 2nd pivot to 1st pivot vector rotated by the 2nd mode of motion (together with the rotated 1st
pivot to atom vectors); and the lanthanide to second pivot vector.

All these vectors are passed into the lib.frame_order.double_rotor.pcs_numeric_int_double_rotor()
function, which passes them to the pcs_pivot_motion_double_rotor() function where they are rotated
and reconstructed into the Ln3+ to atom vectors.