Expansion of the disp_profile_all.py dispersion model super profiling scripts.
The newly added profiling scripts for models 'M61', 'TP02', 'TAP03', and 'MP05' are now included in
the super script to generate statistics for all of these as well.  The nr_iter variable has also
been changed to match the other analytic models, so that the standard deviations are lowered and the
statistics are better.