Updated the dispersion model speed statistics for the disp_spin_speed branch vs. relax-3.2.2.
This now includes the 'NS CPMG 2-site 3D', 'DPL94', and 'NS R1rho 2-site' dispersion models.  The
timings for the single spin analyses are now comparable to the clustered analysis, as the equivalent
of 100 single spins is being used.  The final printout is also in a better format to present for the
relax release messages.

These new results show the insane 160x speed up of the 'DPL94' model!