Created a text file suitable for use as part of the relax release notes.
This contains the statistically averaged profiling information of the speed of the dispersion models
in the disp_spin_speed branch vs. relax-3.2.2.  This file has been created so that it can be used as
part of the release notes for the version of relax that contains the insane speed ups of this
branch.  This file will be updated as new models are profiled and if any more speed ups magically
appear.