Fix for using the old matrix_exponential of m1.
Onye:
- test_korzhnev_2005_15n_sq_data

Is still failing.
That still uses the matrix_exponential_rankN.

There seems to be a problem with matrix_exponential_rankN, when doing
complex numbers.

Maybe the dtype has to get fixed?
Use it as a input argument?

It must be the einsum

Task #7807 (https://gna.org/task/index.php?7807): Speed-up of dispersion models for Clustered analysis.