Math domain fix for ns cpmg 2site expanded.
This is when t108 or t112 is zero, in the multidimensional array, a division error occurs.
The elements are first set to 1.0, to allow for computation.
Then elements are later replaced with 1e100.
Lastly, if the elements are not part of the "True" dispersion point structure, they are cleaned out.

Task #7807 (https://gna.org/task/index.php?7807): Speed-up of dispersion models for Clustered analysis.