Modifications to the dispersion model profiling master script.
The info.print_sys_info() function of the current relax version is being called at the start to show
all information about the current system.  This is useful to know the speed of the machine, the OS,
the Python version and numpy version.  The numpy version is important as future versions might
optimise certain functions that are currently very slow, hence that could be a cause of model speed
differences.

In addition, the path variables path1 and path2 have been renamed to path_new and path_old to make
it clearer which is which.  And the individual profiling scripts are no longer copied to the base
directory of the relax versions being compared, and are run in place.