Added a relaxation dispersion model profiling log file for relax version 3.2.2 vs. 3.2.1.
This is the output from the dispersion model profiling master script.  It shows that the relax 3.2.2
release did not in fact feature "a large speed up of all analytic relaxation dispersion models" as
described in the release notes at https://gna.org/forum/forum.php?forum_id=2465.  For the CPMG
models there is a 1 to 2 times increase in speed.  But for the R1rho models, there is a 1 to 2 times
decrease in speed.