Created a new synthetic CaM data set for the double rotor frame order model.
This is the same as the test_suite/shared_data/frame_order/cam/double_rotor data except that the
angles have been increased from 11.5 and 10.5 degrees to 85.0 and 55.0 for the two torsion angles.
This is to help in debugging the double rotor model as the original test data is too close to the
rigid state to notice certain issues.