Updated the dispersion model profiling comparison of the disp_spin_speed branch vs. relax-3.2.2.
This is for the current revision r24314 and now includes all dispersion models.  This shows the
large speed increases in the numeric and MMQ models recently obtained.

Note that something went wrong with the 'NS CPMG 2-site 3D full' model for the clustered analysis,
most times were around 24 seconds except for the first which was strangely 292 seconds.